Controller of backend is finished

+++ pet_shelter/pet_shelter/src/main/java/com/pet/pet_shelter/DTOs/AdoptionRecord.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AdoptionRecord {
-    int petId;
-    int adopterID;
+    Long petId;
+    Long adopterID;
 }
